More readmes and cleanup
