+++ ATTRIBUTIONS.md
@@ -1,32 +1,20 @@
 # Google Android SDK, Android Support Library, Google Play Services Library
 
 > Copyright (C) 2013 The Android Open Source Project
-> 
+>
 > Licensed under the Apache License, Version 2.0 (the "License");
 > you may not use this file except in compliance with the License.
 > You may obtain a copy of the License at
-> 
+>
 >    http://www.apache.org/licenses/LICENSE-2.0
-> 
+>
 > Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 > See the License for the specific language governing permissions and
 > limitations under the License.
 
 # Google GSON
 
-> Copyright 2008 Google Inc. 
-> 
-> Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
-> 
-> You may obtain a copy of the License at
-> 
->   http://www.apache.org/licenses/LICENSE-2.0
-> 
-> Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
-
-# Picasso (used in samples)
-
-> Copyright 2013 Square, Inc.
+> Copyright 2008 Google Inc.
 >
 > Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
 >
