+++ samples/AdvancedPushNotifications/README
@@ -1,14 +1,36 @@
-Eclipse ADT build instructions
-==================================================
--Include SwrveSDK project in workspace.
--Include SwrveSDKDemo project in workspace.
--Replace YOUR_APP_ID in MainActivity.java with your app id.
--Replace YOUR_API_KEY in MainActivity.java with your api key.
--Run SwrveSDKDemo app normally.
+Swrve SDK Advanced Push Notifications Sample
+--------------------------------------------
+Example of how to create an advanced UI for your push notifications, taking advantage of Android's latest features and Swrve's custom payloads for extra data.
+
+This example extends and replaces the SwrveGcmIntentService to provide a big notification style and grouping or a custom XML layout for your notifications:
+- [BigNotificationSwrveGcmIntentService](src/main/java/com/swrve/sdk/sample/BigNotificationSwrveGcmIntentService.java)
+- [CustomLayoutSwrveGcmIntentService](src/main/java/com/swrve/sdk/sample/CustomLayoutSwrveGcmIntentService.java)
+
+
+For more information on Android's notification features have a look at the following links:
+- http://developer.android.com/guide/topics/ui/notifiers/notifications.html
+- http://developer.android.com/training/notify-user/expanded.html
+- http://developer.android.com/training/wearables/notifications/stacks.html
+- http://developer.android.com/guide/topics/ui/notifiers/notifications.html
 
 Android Studio build instructions
-==================================================
--Import SwrveSDKDemo.
--Replace YOUR_APP_ID in MainActivity.java with your app id.
--Replace YOUR_API_KEY in MainActivity.java with your api key.
--Run SwrveSDKDemo app normally.
+---------------------------------
+- Import AdvancedPushNotifications.
+- Replace YOUR_APP_ID in SampleApplication.java with your Swrve app ID.
+- Replace YOUR_API_KEY in SampleApplication.java with your Swrve API key.
+- Replace YOUR_SENDER_ID in SampleApplication.java with your Google Cloud Messaging Sender ID if you want to do push notifications.
+- Run AdvancedPushNotifications app normally.
+
+Attributions
+------------
+# Picasso
+
+> Copyright 2013 Square, Inc.
+>
+> Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
+>
+> You may obtain a copy of the License at
+>
+>   http://www.apache.org/licenses/LICENSE-2.0
+>
+> Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
