+++ samples/MessageCenter/README
@@ -1,14 +1,14 @@
-Eclipse ADT build instructions
-==================================================
--Include SwrveSDK project in workspace.
--Include SwrveSDKDemo project in workspace.
--Replace YOUR_APP_ID in MainActivity.java with your app id.
--Replace YOUR_API_KEY in MainActivity.java with your api key.
--Run SwrveSDKDemo app normally.
+Swrve SDK Message Center Sample
+-------------------------------
+Example of how to use the [Message Center API](http://docs.swrve.com/developer-documentation/api-guides/swrve-message-center-api-guide/) to build an inbox-style UI.
+
+The following class contains an adaptor and fragment to render the campaigns:
+- [MessageCenterFragment](src/main/java/com/swrve/sdk/sample/MessageCenterFragment.java)
 
 Android Studio build instructions
-==================================================
--Import SwrveSDKDemo.
--Replace YOUR_APP_ID in MainActivity.java with your app id.
--Replace YOUR_API_KEY in MainActivity.java with your api key.
--Run SwrveSDKDemo app normally.
+---------------------------------
+- Import MessageCenter.
+- Replace YOUR_APP_ID in SampleApplication.java with your Swrve app ID.
+- Replace YOUR_API_KEY in SampleApplication.java with your Swrve API key.
+- Replace YOUR_SENDER_ID in SampleApplication.java with your Google Cloud Messaging Sender ID if you want to do push notifications.
+- Run MessageCenter app normally.
