+++ samples/MultipleGCMProviders/proguard-rules.pro
