+++ samples/MultipleGCMProviders/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     <TextView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="Advanced Push Notifications sample.\n\n Have a look at the code!"
+        android:text="Multiple GCM Push Providers sample.\n\n Have a look at the code!"
         android:gravity="center"
         android:textSize="30dp" />
 </LinearLayout>
