+++ samples/MultipleGCMProviders/src/main/res/mipmap-hdpi/ic_launcher_material.png
