+++ samples/MultipleGCMProviders/src/main/res/mipmap-mdpi/ic_launcher_material.png
