+++ samples/MultipleGCMProviders/src/main/res/mipmap-xhdpi/ic_launcher_material.png
