+++ samples/MultipleGCMProviders/src/main/res/mipmap-xxhdpi/ic_launcher_material.png
