+++ samples/MultipleGCMProviders/src/main/res/mipmap-xxxhdpi/ic_launcher.png
