+++ samples/MultipleGCMProviders/src/main/res/mipmap/ic_launcher.png
