+++ samples/MultipleGCMProviders/src/main/res/mipmap/ic_launcher_material.png
