+++ samples/MultipleGCMProviders/src/main/res/values/dimens.xml
