+++ samples/MultipleGCMProviders/src/main/res/values/strings.xml
