+++ samples/MultipleGCMProviders/src/main/res/values/styles.xml
