+++ samples/README
@@ -1,0 +1,6 @@
+Swrve SDK Samples
+-----------------
+Examples of advanced usage of the Swrve SDK:
+- [Integrate push notifications with another provider](MultipleGCMProviders/)
+- [Message Center API](MessageCenter/)
+- [Advanced Push Notifications UI](AdvancedPushNotifications/)
